LayoutTests/platform/chromium/accessibility/aria-tree-index-of-items-expected.txt

Animals
Birds
Cats
Siamese
Tabby
This tests that the index attribute is correctly returned for all tree items, regardless of their hierarchical level.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axtree.selectedRowAtIndex(0).indexInTable() is 0
PASS axtree.selectedRowAtIndex(1).indexInTable() is 1
PASS axtree.selectedRowAtIndex(2).indexInTable() is 2
PASS axtree.selectedRowAtIndex(3).indexInTable() is 3
PASS axtree.selectedRowAtIndex(4).indexInTable() is 4
PASS successfullyParsed is true

TEST COMPLETE
